% Generated by build script
\name{wb_countries}
\alias{wb_countries}
\title{List World Bank countries and regions}
\usage{
wb_countries()
}
\arguments{
\item{per_page}{Results per page (default 300)}
}
\value{tibble: iso3, iso2, name, region, income_level, capital,}
\description{List World Bank countries and regions}
